Extract `flat` module
