+++ src/flat.rs
@@ -1,0 +1,30 @@
+use failure::Error;
+use flat_projection::{FlatProjection, FlatPoint};
+use ord_subset::OrdSubsetIterExt;
+
+use crate::Point;
+use crate::parallel::*;
+
+/// Projects all geographic points onto a flat surface for faster geodesic calculation
+///
+pub fn to_flat_points<T: Point>(points: &[T]) -> Vec<FlatPoint<f64>> {
+    let center = points.center_lat().unwrap();
+    let proj = FlatProjection::new(center);
+
+    opt_par_iter(points)
+        .map(|fix| proj.project(fix.longitude(), fix.latitude()))
+        .collect()
+}
+
+trait CenterLatitude {
+    fn center_lat(self: &Self) -> Option<f64>;
+}
+
+impl<T: Point> CenterLatitude for [T] {
+    fn center_lat(self: &Self) -> Option<f64> {
+        let lat_min = self.iter().map(|fix| fix.latitude()).ord_subset_min()?;
+        let lat_max = self.iter().map(|fix| fix.latitude()).ord_subset_max()?;
+
+        Some((lat_min + lat_max) / 2.)
+    }
+}
